+++ packages/common/test/Evolu/Timestamp.test.ts
@@ -2,26 +2,26 @@ import SQLite from "better-sqlite3";
 import { describe, expect, test } from "vitest";
 import { defaultDbConfig } from "../../src/Evolu/Db.js";
 import {
-  BinaryTimestamp,
   Counter,
   Millis,
   NodeId,
   Timestamp,
+  TimestampBytes,
   TimestampConfigDep,
   TimestampCounterOverflowError,
   TimestampDriftError,
   TimestampTimeOutOfRangeError,
-  binaryTimestampToTimestamp,
   createInitialTimestamp,
   createTimestamp,
   maxCounter,
   maxMillis,
   minCounter,
   minMillis,
-  orderBinaryTimestamp,
+  orderTimestampBytes,
   receiveTimestamp,
   sendTimestamp,
-  timestampToBinaryTimestamp,
+  timestampBytesToTimestamp,
+  timestampToTimestampBytes,
   timestampToTimestampString,
 } from "../../src/Evolu/Timestamp.js";
 import { increment } from "../../src/Number.js";
@@ -320,51 +320,51 @@ describe("receiveTimestamp", () => {
     });
   });
 
-  test("timestampToBinaryTimestamp/binaryTimestampToTimestamp", () => {
-    const decodeFromEncoded = (t: BinaryTimestamp) =>
-      binaryTimestampToTimestamp(t);
+  test("timestampToTimestampBytes/timestampBytesToTimestamp", () => {
+    const decodeFromEncoded = (t: TimestampBytes) =>
+      timestampBytesToTimestamp(t);
 
     const t = createTimestamp();
-    expect(t).toStrictEqual(decodeFromEncoded(timestampToBinaryTimestamp(t)));
+    expect(t).toStrictEqual(decodeFromEncoded(timestampToTimestampBytes(t)));
 
     const lastSafeTimestampEncodedDecoded = decodeFromEncoded(
-      timestampToBinaryTimestamp(createTimestamp({ millis: maxMillis })),
+      timestampToTimestampBytes(createTimestamp({ millis: maxMillis })),
     );
     expect(lastSafeTimestampEncodedDecoded.millis).toBe(maxMillis);
 
-    const t1 = timestampToBinaryTimestamp(
+    const t1 = timestampToTimestampBytes(
       createTimestamp({ millis: minMillis }),
     );
-    const t2 = timestampToBinaryTimestamp(
+    const t2 = timestampToTimestampBytes(
       createTimestamp({
         millis: Millis.orThrow(increment(minMillis)),
       }),
     );
-    expect(orderBinaryTimestamp(t1, t2)).toBe(-1);
-    expect(orderBinaryTimestamp(t2, t1)).toBe(1);
-    expect(orderBinaryTimestamp(t1, t1)).toBe(0);
+    expect(orderTimestampBytes(t1, t2)).toBe(-1);
+    expect(orderTimestampBytes(t2, t1)).toBe(1);
+    expect(orderTimestampBytes(t1, t1)).toBe(0);
 
-    const t3 = timestampToBinaryTimestamp(
+    const t3 = timestampToTimestampBytes(
       createTimestamp({ counter: minCounter }),
     );
-    const t4 = timestampToBinaryTimestamp(
+    const t4 = timestampToTimestampBytes(
       createTimestamp({
         counter: Counter.orThrow(increment(minCounter)),
       }),
     );
-    expect(orderBinaryTimestamp(t3, t4)).toBe(-1);
-    expect(orderBinaryTimestamp(t4, t3)).toBe(1);
-    expect(orderBinaryTimestamp(t3, t3)).toBe(0);
+    expect(orderTimestampBytes(t3, t4)).toBe(-1);
+    expect(orderTimestampBytes(t4, t3)).toBe(1);
+    expect(orderTimestampBytes(t3, t3)).toBe(0);
 
-    const t5 = timestampToBinaryTimestamp(
+    const t5 = timestampToTimestampBytes(
       createTimestamp({ nodeId: "0000000000000000" as NodeId }),
     );
-    const t6 = timestampToBinaryTimestamp(
+    const t6 = timestampToTimestampBytes(
       createTimestamp({ nodeId: "0000000000000001" as NodeId }),
     );
-    expect(orderBinaryTimestamp(t5, t6)).toBe(-1);
-    expect(orderBinaryTimestamp(t6, t5)).toBe(1);
-    expect(orderBinaryTimestamp(t5, t5)).toBe(0);
+    expect(orderTimestampBytes(t5, t6)).toBe(-1);
+    expect(orderTimestampBytes(t6, t5)).toBe(1);
+    expect(orderTimestampBytes(t5, t5)).toBe(0);
 
     const randomMillis = new Set<Millis>();
     Array.from({ length: 1000 }).forEach(() => {
@@ -373,13 +373,13 @@ describe("receiveTimestamp", () => {
 
     const sortedMillis = [...randomMillis].toSorted(orderNumber);
 
-    const randomBinaryTimestamps = [...randomMillis]
+    const randomTimestampsBytes = [...randomMillis]
       .map((millis) => createTimestamp({ millis }))
-      .map(timestampToBinaryTimestamp);
+      .map(timestampToTimestampBytes);
 
     expect(
-      randomBinaryTimestamps
-        .toSorted(orderBinaryTimestamp)
+      randomTimestampsBytes
+        .toSorted(orderTimestampBytes)
         .map(decodeFromEncoded)
         .map((a) => a.millis),
     ).toEqual(sortedMillis);
@@ -395,11 +395,11 @@ describe("receiveTimestamp", () => {
     ).run();
 
     const insertTimestamp = db.prepare(`insert into Message (t) values (@t)`);
-    randomBinaryTimestamps.forEach((t) => {
+    randomTimestampsBytes.forEach((t) => {
       insertTimestamp.run({ t });
     });
     const sqliteMillis = db
-      .prepare<[], { t: BinaryTimestamp }>(`select t from Message order by t`)
+      .prepare<[], { t: TimestampBytes }>(`select t from Message order by t`)
       .all()
       .map((a) => decodeFromEncoded(a.t).millis);
     expect(sqliteMillis).toEqual(sortedMillis);